Refactor transaction handling and Telegram bot logic
Updated `TransactionId` type to `long` across relevant classes to support larger IDs. Enhanced `FinanceControlDefineCategoryHandler` to include category retrieval and transaction amount printing via Telegram bot. Added `/definircategoria` command handling in `FinanceControlMessageReceivedHandler`. Improved `FinancialControlCreateTransactionMapper` by trimming and uppercasing `CreditCard` field. Refactored `FinancialControlHelper` to include new inline keyboard buttons and added `PrintNewCategory` method. Made `StringHelper` static with regex-based value extraction. Adjusted repository interfaces and methods to accommodate `long` transaction IDs.

+++ TelegramBotManager/Application/Features/FinanceControlDefineCategory/FinanceControlDefineCategoryCommand.cs
@@ -4,13 +4,13 @@ namespace TelegramBotManager.Application.Features.FinanceControlDefineCategory;
 
 public class FinanceControlDefineCategoryCommand : IRequest<Unit>
 {
-    public FinanceControlDefineCategoryCommand(int transactionId, long? categoryId = null)
+    public FinanceControlDefineCategoryCommand(long transactionId, long? categoryId = null)
     {
         TransactionId = transactionId;
         CategoryId = categoryId;
     }
 
-    public int TransactionId { get; set; }
+    public long TransactionId { get; set; }
 
     public long? CategoryId { get; set; }
 }
