+++ TelegramBotManager/Domain/Interfaces/ITransactionRepository.cs
@@ -4,7 +4,7 @@ namespace TelegramBotManager.Domain.Interfaces;
 
 public interface ITransactionRepository
 {
-    Task<transaction> GetTransactionById(int transactionId, CancellationToken cancellationToken);
+    Task<transaction> GetTransactionById(long transactionId, CancellationToken cancellationToken);
     Task<List<transaction>> GetTransactions(CancellationToken cancellationToken);
     Task<transaction> SaveAsync(transaction transaction, CancellationToken cancellationToken);
     Task<decimal> GetAmmountOfMonth(transaction transaction, CancellationToken cancellationToken, bool filterByCategory = false);
